# api/openapi.yaml
openapi: 3.0.3
info:
  title: Monelog API
  version: 1.0.0
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Transaction:
      type: object
      properties:
        id: { type: integer, format: int64 }
        user_id: { type: integer, format: int64 }
        category_id: { type: integer, format: int64 }
        amount: { type: integer, format: int64 }
        note: { type: string }
        date: { type: string, format: date }
        created_at: { type: string, format: date-time }
    CreateTransactionRequest:
      type: object
      properties:
        category_id: { type: integer, format: int64 }
        amount: { type: integer, format: int64 }
        date: { type: string, format: date }
        note: { type: string }
      required: [category_id, amount, date]
    UpdateTransactionRequest:
      type: object
      properties:
        category_id: { type: integer, format: int64 }
        amount: { type: integer, format: int64 }
        date: { type: string, format: date }
        note: { type: string }
      required: [category_id, amount, date]
...
    TransactionDetail:
      type: object
      properties:
        date: { type: string, format: date }
        category_name: { type: string }
        category_type: { type: string }
        note: { type: string }
        total_income: { type: integer, format: int64 }
        total_expense: { type: integer, format: int64 }
        balance: { type: integer, format: int64 }
security:
  - BearerAuth: []
paths:
...
  /transactions:
    post:
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTransactionRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
#    get:
#      operationId: getTransactions
#      parameters:
#        - name: from
#          in: query
#          required: false
#          schema: { type: string, format: date }
#        - name: to
#          in: query
#          required: false
#          schema: { type: string, format: date }
#      responses:
#        '200':
#          description: OK
#          content:
#            application/json:
#              schema:
#                type: array
#                items: { $ref: '#/components/schemas/Transaction' }
  /transactions/{id}:
    put:
      operationId: updateTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTransactionRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
    delete:
      operationId: softDeleteTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: Deleted }
        '404': { description: Not found }
...
  /reports/date-range:
    get:
      operationId: getTransactionsBetweenDatesWithNote
      security:
        - BearerAuth: [ ]
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TransactionDetail' }